REF Remove ScatteringEntry out of __init__
To maintain consistency, we should reserve `__init__` modules for just
importing from other packages instead of defining actual classes and
methods in there. The entry point classes (`ScatteringEntry*D`) are
therefore moved to `entry` modules that are then imported directly.
