+++ kymatio/scattering1d/__init__.py
@@ -1,3 +1,3 @@
-from .frontend import ScatteringEntry1D
+from .frontend.entry import ScatteringEntry1D
 
 __all__ = ['ScatteringEntry1D']
