+++ kymatio/scattering2d/__init__.py
@@ -1,4 +1,4 @@
-from .frontend import ScatteringEntry2D
+from .frontend.entry import ScatteringEntry2D
 
 
 __all__ = ['ScatteringEntry2D']
