+++ kymatio/scattering3d/__init__.py
@@ -1,4 +1,4 @@
-from .frontend import HarmonicScatteringEntry3D
+from .frontend.entry import HarmonicScatteringEntry3D
 
 
 __all__ = ['HarmonicScatteringEntry3D']
