+++ kymatio/scattering3d/frontend/entry.py
@@ -1,0 +1,11 @@
+from ...frontend.entry import ScatteringEntry
+
+
+class HarmonicScatteringEntry3D(ScatteringEntry):
+    def __init__(self, *args, **kwargs):
+        super().__init__(name='harmonic 3D',
+                         class_name='scattering3d',
+                         *args, **kwargs)
+
+
+__all__ = ['HarmonicScatteringEntry3D']
